+++ app/src/main/res/layout/fragment_change_password.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ChangePasswordFragment">
+    tools:context="com.alt.letseatingtime_android.ui.fragment.pwchange.ChangePasswordFragment">
 
     <ImageView
         android:id="@+id/iv_eating_food"
