+++ app/src/main/res/navigation/nav_login.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/loginFragment"
-        android:name="com.alt.letseatingtime.LoginFragment"
+        android:name="com.alt.letseatingtime_android.ui.fragment.login.LoginFragment"
         android:label="fragment_login"
         tools:layout="@layout/fragment_login" >
         <action
@@ -19,7 +19,7 @@
     </fragment>
     <fragment
         android:id="@+id/signupFragment42"
-        android:name="com.alt.letseatingtime_android.ui.fragment.SignupFragment4"
+        android:name="com.alt.letseatingtime_android.ui.fragment.signup.SignupFragment4"
         android:label="SignupFragment4" >
         <action
             android:id="@+id/action_signupFragment42_to_signupFragment32"
@@ -30,7 +30,7 @@
     </fragment>
     <fragment
         android:id="@+id/signupFragment12"
-        android:name="com.alt.letseatingtime_android.ui.fragment.SignupFragment1"
+        android:name="com.alt.letseatingtime_android.ui.fragment.signup.SignupFragment1"
         android:label="SignupFragment1" >
         <action
             android:id="@+id/action_signupFragment12_to_signupFragment22"
@@ -41,7 +41,7 @@
     </fragment>
     <fragment
         android:id="@+id/signupFragment22"
-        android:name="com.alt.letseatingtime_android.ui.fragment.SignupFragment2"
+        android:name="com.alt.letseatingtime_android.ui.fragment.signup.SignupFragment2"
         android:label="SignupFragment2" >
         <action
             android:id="@+id/action_signupFragment22_to_signupFragment32"
@@ -52,7 +52,7 @@
     </fragment>
     <fragment
         android:id="@+id/signupFragment32"
-        android:name="com.alt.letseatingtime_android.ui.fragment.SignupFragment3"
+        android:name="com.alt.letseatingtime_android.ui.fragment.signup.SignupFragment3"
         android:label="SignupFragment3" >
         <action
             android:id="@+id/action_signupFragment32_to_signupFragment42"
@@ -63,7 +63,7 @@
     </fragment>
     <fragment
         android:id="@+id/changePasswordFragment"
-        android:name="com.alt.letseatingtime.ChangePasswordFragment"
+        android:name="com.alt.letseatingtime_android.ui.fragment.pwchange.ChangePasswordFragment"
         android:label="fragment_change_password"
         tools:layout="@layout/fragment_change_password" >
         <action
